Make sure nb_workers_stereo is int

+++ s2p/__init__.py
@@ -614,7 +614,7 @@ def main(user_cfg, start_from=0):
             # Set the number of stereo workers to 2/3 of the number of cores by default, divided
             # by a certain amount depending on the tile_size this should be a generally safe number of workers.
             divider = (cfg['tile_size'] / 800.0) * (cfg['tile_size'] / 800.0)
-            nb_workers_stereo = min(nb_workers, max(1, int(nb_workers / 2.0) / divider))
+            nb_workers_stereo = int(min(nb_workers, max(1, int(nb_workers / 2.0) / divider)))
         try:
             print(f'4) running stereo matching using {nb_workers_stereo} workers...')
             parallel.launch_calls(stereo_matching, tiles_pairs, nb_workers_stereo,
